+++ lib/postgrex/utils.ex
@@ -1,13 +1,6 @@
 defmodule Postgrex.Utils do
   @moduledoc false
 
-  import Postgrex.BinaryUtils
-
-  def encode_param(<<-1 :: int32>>),
-    do: <<-1 :: int32>>
-  def encode_param(param),
-    do: [<<IO.iodata_length(param) :: int32>>, param]
-
   def error(error, s) do
     reply(error, s)
     {:stop, error, s}